+++ src/main/kotlin/org/jetbrains/research/ml/ast/cli/Runner.kt
@@ -5,25 +5,19 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ApplicationStarter
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.projectRoots.Sdk
-import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import com.jetbrains.python.PythonFileType
-import com.jetbrains.python.configuration.PyConfigurableInterpreterList
-import com.jetbrains.python.sdk.PythonSdkType
-import com.jetbrains.python.sdk.configuration.PyProjectVirtualEnvConfiguration
 import com.xenomachina.argparser.ArgParser
-import org.jetbrains.annotations.NotNull
 import org.jetbrains.research.ml.ast.transformations.Transformation
 import org.jetbrains.research.ml.ast.util.*
+import org.jetbrains.research.ml.ast.util.sdk.setSdkToProject
 import java.io.File
 import java.nio.file.Paths
 import kotlin.random.Random.Default.nextDouble
 import kotlin.system.exitProcess
 
-// TODO: test it with different transformations sets
 object Runner : ApplicationStarter {
     private lateinit var inputDir: String
     private lateinit var outputDir: String
@@ -52,14 +46,6 @@ object Runner : ApplicationStarter {
             "--yaml_path",
             help = "YAML config path"
         )
-    }
-
-    private fun getTmpProjectDir(toCreateFolder: Boolean = true): String {
-        val path = "${System.getProperty("java.io.tmpdir").removeSuffix("/")}/astTransformationsTmp"
-        if (toCreateFolder) {
-            createFolder(path)
-        }
-        return path
     }
 
     // Filter transformations according to the probability
@@ -91,45 +77,10 @@ object Runner : ApplicationStarter {
         ApplicationManager.getApplication().invokeAndWait {
             ApplicationManager.getApplication().runWriteAction {
                 transformations.forEach {
-                    println(it.key)
                     it.forwardApply(this)
                 }
             }
         }
-    }
-
-    private fun createSdk(project: Project, baseSdk: Sdk): Sdk {
-        var sdk: Sdk? = null
-        ApplicationManager.getApplication().invokeAndWait {
-            sdk = PyProjectVirtualEnvConfiguration.createVirtualEnvSynchronously(
-                baseSdk = baseSdk,
-                existingSdks = listOf(baseSdk),
-                venvRoot = getTmpProjectDir(),
-                projectBasePath = project.basePath,
-                project = project,
-                module = null
-            )
-        }
-        return sdk ?: error("Internal error: SDK for temp project was not created")
-    }
-
-    private fun addPyFileToProject(
-        projectPath: String,
-        fileName: String,
-        fileContext: String = ""
-    ): File {
-        val filePath = "$projectPath/$fileName"
-        val file = File(filePath)
-        file.createNewFile()
-        file.writeText(fileContext)
-        return file
-    }
-
-    private fun createBaseSdk(project: Project): Sdk {
-        val myInterpreterList = PyConfigurableInterpreterList.getInstance(project)
-        val myProjectSdksModel = myInterpreterList.model
-        val pySdkType = PythonSdkType.getInstance()
-        return myProjectSdksModel.createSdk(pySdkType, "/usr/bin/python3")
     }
 
     override fun main(args: List<String>) {
@@ -141,15 +92,8 @@ object Runner : ApplicationStarter {
             }
 
             project = ProjectUtil.openOrImport(getTmpProjectDir(), null, true)
-
             project?.let {
-                val baseSdk = createBaseSdk(it)
-
-                val projectManager = ProjectRootManager.getInstance(it)
-                val sdk = createSdk(it, baseSdk)
-                val sdkConfigurer = SdkConfigurer(it, projectManager)
-                sdkConfigurer.setProjectSdk(sdk)
-
+                setSdkToProject(it, getTmpProjectDir())
                 val psiManager = PsiManager.getInstance(it)
                 createFolder(outputDir)
                 val config = Configuration.parseYamlConfig(getContentFromFile(yaml_config))
