+++ src/main/kotlin/org/jetbrains/research/ml/ast/util/sdk/SdkConfigurer.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.ml.ast.util
+package org.jetbrains.research.ml.ast.util.sdk
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
