--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/impure/in_1_global_variable.py
@@ -1,13 +1,0 @@
-y = 5
-
-def impure_foo(s):
-    if s:
-        x = 1
-        y = x
-        print(y)
-        print('foo')
-    else:
-        x = 1
-        y = x
-        print(y)
-        print('bar')
