--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/impure/out_1.py
@@ -1,10 +1,0 @@
-y = 5
-
-def impure_foo(s):
-    x = 1
-    y = x
-    print(y)
-    if s:
-        print('foo')
-    else:
-        print('bar')
