--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/impure/out_2.py
@@ -1,11 +1,0 @@
-import random
-
-def impure_foo(s, y):
-    i = random.randint(1, 100)
-    if i >= 35:
-        s = i
-    else:
-        s = 10
-    x = 1
-    y = x
-    print(y)
