--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_1_simple_prefix.py
@@ -1,12 +1,0 @@
-s = input()
-
-if s:
-    x = 1
-    y = x
-    print(y)
-    print('foo')
-else:
-    x = 1
-    y = x
-    print(y)
-    print('bar')
