--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_2_simple_suffix.py
@@ -1,12 +1,0 @@
-s = input()
-
-if s:
-    print('foo')
-    x = 1
-    y = x
-    print(y)
-else:
-    print('bar')
-    x = 1
-    y = x
-    print(y)
