--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_3_simple_both.py
@@ -1,16 +1,0 @@
-s = input()
-
-if s:
-    a = 2
-    b = a + a
-    print('foo')
-    x = 1
-    y = x
-    print(y)
-else:
-    a = 2
-    b = a + a
-    print('bar')
-    x = 1
-    y = x
-    print(y)
