--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_4_if_elif_else.py
@@ -1,24 +1,0 @@
-s = input()
-z = input()
-
-if s:
-    a = 2
-    b = a + a
-    print('foo')
-    x = 1
-    y = x
-    print(y)
-elif z:
-    a = 2
-    b = a + a
-    print('bar')
-    x = 1
-    y = x
-    print(y)
-else:
-    a = 2
-    b = a + a
-    print('baz')
-    x = 1
-    y = x
-    print(y)
