--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_5_remove_whole_if.py
@@ -1,14 +1,0 @@
-s = input()
-
-if s:
-    a = 2
-    b = a + a
-    x = 1
-    y = x
-    print(y)
-else:
-    a = 2
-    b = a + a
-    x = 1
-    y = x
-    print(y)
