--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_6_remove_if_before_else.py
@@ -1,15 +1,0 @@
-s = input()
-
-if s:
-    a = 2
-    b = a + a
-    x = 1
-    y = x
-    print(y)
-else:
-    a = 2
-    b = a + a
-    x = 1
-    y = x
-    print(y)
-    print('bar')
