--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_7_remove_elif.py
@@ -1,22 +1,0 @@
-s = input()
-
-if s:
-    a = 2
-    b = a + a
-    print('foo')
-    x = 1
-    y = x
-    print(y)
-elif s:
-    a = 2
-    b = a + a
-    x = 1
-    y = x
-    print(y)
-else:
-    a = 2
-    b = a + a
-    print('baz')
-    x = 1
-    y = x
-    print(y)
