--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_8_foo.py
@@ -1,11 +1,0 @@
-def pure_foo(s, y):
-    if s:
-        x = 1
-        y = x
-        print(y)
-        print('foo')
-    else:
-        x = 1
-        y = x
-        print(y)
-        print('bar')
