--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/in_9_nested.py
@@ -1,17 +1,0 @@
-s = input()
-
-if True:
-    if s:
-        a = 2
-        b = a + a
-        print('foo')
-        x = 1
-        y = x
-        print(y)
-    else:
-        a = 2
-        b = a + a
-        print('bar')
-        x = 1
-        y = x
-        print(y)
