--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_1.py
@@ -1,9 +1,0 @@
-s = input()
-
-x = 1
-y = x
-print(y)
-if s:
-    print('foo')
-else:
-    print('bar')
