--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_10.py
@@ -1,12 +1,0 @@
-s = input()
-
-if s:
-    print('foo')
-    a = 2
-    b = a + a
-    print('bar')
-else:
-    print('foo1')
-    a = 2
-    b = a + a
-    print('bar1')
