--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_3.py
@@ -1,11 +1,0 @@
-s = input()
-
-a = 2
-b = a + a
-if s:
-    print('foo')
-else:
-    print('bar')
-x = 1
-y = x
-print(y)
