--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_4.py
@@ -1,14 +1,0 @@
-s = input()
-z = input()
-
-a = 2
-b = a + a
-if s:
-    print('foo')
-elif z:
-    print('bar')
-else:
-    print('baz')
-x = 1
-y = x
-print(y)
