--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_5.py
@@ -1,7 +1,0 @@
-s = input()
-
-a = 2
-b = a + a
-x = 1
-y = x
-print(y)
