--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_6.py
@@ -1,9 +1,0 @@
-s = input()
-
-a = 2
-b = a + a
-x = 1
-y = x
-print(y)
-if not ((s)):
-    print('bar')
