--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_7.py
@@ -1,11 +1,0 @@
-s = input()
-
-a = 2
-b = a + a
-if s:
-    print('foo')
-else:
-    print('baz')
-x = 1
-y = x
-print(y)
