--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_8.py
@@ -1,8 +1,0 @@
-def pure_foo(s, y):
-    x = 1
-    y = x
-    print(y)
-    if s:
-        print('foo')
-    else:
-        print('bar')
