--- src/test/resources/org/jetbrains/research/ml/ast/transformations/if_redundant_lines_removal/data/pure/out_9.py
@@ -1,12 +1,0 @@
-s = input()
-
-if True:
-    a = 2
-    b = a + a
-    if s:
-        print('foo')
-    else:
-        print('bar')
-    x = 1
-    y = x
-    print(y)
